Icons and mockups oh my
The brain is melting and impatient with the computer and the wrist, so sleep is impending. But we have a little of almost everything. Just gotta round out the details like having more than one tower (because untangling the trebuchet ended up being more than late night me was up for, past the legs).